// Frontend/src/app/model/x_his_auftrag.ts
export class XHisAuftrag {
    name: string;
    EO_ID: string;
    ERROR_ID: number;
    ID: string;
    KDNR: string;
    ORDER_ID_BSSOE: string;
    ORDER_ID_CRM: string;
    ORDER_ID_DKK: string;
    ORDER_ID_MISC: string;
    ORDER_ID_SMF: string;
    ORDER_ID_SNR: string;
    SEQ_ID: number;
    SO_ID: string;
    SO_TYPE_ID: string;
    STATUS: string;
    TO_NR: string;
    TS_ABORTED: string;
    TS_COMMITED: string;
    TS_COMPLETED: string;
    TS_CREATED: string;
    TS_EXECUTED: string;
    TS_HISTORY: string;
    TS_LAST_UPDATE: string;
    TS_SAVED: string;
    URSPRUNGS_KWT: string;
}